+++ src/main/kotlin/app/revanced/patches/twitter/layout/hideviews/annotations/HideViewsCompatibility.kt
@@ -5,7 +5,7 @@ import app.revanced.patcher.annotation.Package
 
 @Compatibility(
     [Package(
-        "com.twitter.android", arrayOf("9.69.1-release.0")
+        "com.twitter.android", arrayOf("9.69.1-release.0", "9.71.0-release.0")
     )]
 )
 @Target(AnnotationTarget.CLASS)