+++ src/main/kotlin/app/revanced/patches/youtube/layout/returnyoutubedislike/fingerprints/ShortsTextComponentParentFingerprint.kt
@@ -1,0 +1,25 @@
+package app.revanced.patches.youtube.layout.returnyoutubedislike.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.AccessFlags
+import org.jf.dexlib2.Opcode
+
+object ShortsTextComponentParentFingerprint : MethodFingerprint(
+    returnType = "V",
+    access = AccessFlags.PROTECTED or AccessFlags.FINAL,
+    parameters = listOf("L", "L"),
+    opcodes = listOf(
+        Opcode.IF_EQZ,
+        Opcode.CONST_4,
+        Opcode.IF_EQ,
+        Opcode.CONST_4,
+        Opcode.IF_EQ,
+        Opcode.RETURN_VOID,
+        Opcode.IGET_OBJECT,
+        Opcode.CHECK_CAST,
+        Opcode.IGET_BOOLEAN,
+        Opcode.IF_EQZ,
+        Opcode.INVOKE_STATIC
+    )
+)